+++ ui/CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to the **Prowler UI** are documented in this file.
 - Navigation progress bar for page transitions using Next.js `onRouterTransitionStart` [(#9465)](https://github.com/prowler-cloud/prowler/pull/9465)
 - Finding Severity Over Time chart component to Overview page [(#9405)](https://github.com/prowler-cloud/prowler/pull/9405)
 - Attack Surface component to Overview page [(#9412)](https://github.com/prowler-cloud/prowler/pull/9412)
-- Resource details (partition, datail and metadata) [(#9515)](https://github.com/prowler-cloud/prowler/pull/9515)
+- Resource details (partition, details and metadata) [(#9515)](https://github.com/prowler-cloud/prowler/pull/9515)
 
 ### 🔄 Changed